refactor(admin/space): 불필요한 pricerange 모듈 제거 및 패키지 병합
