+++ backend/src/main/java/com/coliving/admin/space/adapter/in/web/dto/req/CreatePriceRangeRequestDto.java
@@ -1,4 +1,4 @@
-package com.coliving.admin.pricerange.adapter.in.web.dto.req;
+package com.coliving.admin.space.adapter.in.web.dto.req;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class UpdatePriceRangeRequestDto {
+public class CreatePriceRangeRequestDto {
     @NotBlank(message = "라벨은 필수입니다.")
     private String label;
     private Integer minRent;
