+++ backend/src/main/java/com/coliving/admin/space/application/port/in/AdminPriceRangeUseCase.java
@@ -1,8 +1,8 @@
-package com.coliving.admin.pricerange.application.port.in;
+package com.coliving.admin.space.application.port.in;
 
-import com.coliving.admin.pricerange.application.command.CreatePriceRangeCommand;
-import com.coliving.admin.pricerange.application.command.UpdatePriceRangeCommand;
-import com.coliving.admin.pricerange.application.result.AdminPriceRangeResult;
+import com.coliving.admin.space.application.command.CreatePriceRangeCommand;
+import com.coliving.admin.space.application.command.UpdatePriceRangeCommand;
+import com.coliving.admin.space.application.result.AdminPriceRangeResult;
 
 import java.util.List;
