+++ backend/src/main/java/com/coliving/user/room/adapter/in/web/dto/PriceRangeResponseDto.java
@@ -1,20 +1,20 @@
 package com.coliving.user.room.adapter.in.web.dto;
 
-import com.coliving.admin.pricerange.adapter.out.jpa.PriceRangePresetEntity;
+import com.coliving.admin.space.adapter.out.jpa.PriceRangePresetEntity;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class PublicPriceRangeResponseDto {
+public class PriceRangeResponseDto {
     private final Long priceRangePresetId;
     private final String label;
     private final Integer minRent;
     private final Integer maxRent;
     private final Integer sortOrder;
 
-    public static PublicPriceRangeResponseDto from(PriceRangePresetEntity entity) {
-        return PublicPriceRangeResponseDto.builder()
+    public static PriceRangeResponseDto from(PriceRangePresetEntity entity) {
+        return PriceRangeResponseDto.builder()
                 .priceRangePresetId(entity.getId())
                 .label(entity.getLabel())
                 .minRent(entity.getMinRent())
